HDR assembly changed
old: PSL is produced followed by HDR assembly.
new: first HDR is assembled and afterwards the picture is converted to a PSL.

The reason for this change is in disexcitation of the imageplate by multible read in.
The intensity of every pixel is attenuated by an arbitrary factor. So this process is linear.
The PSL conversion in contast is non linear. So first you have to assemble the multible read-ins to a HDR and do afterwards the PSL conversion.

For more information read:
http://www.ndt.net/article/apcndt2013/papers/059.pdf
https://www.jstage.jst.go.jp/article/jnst1964/34/5/34_5_461/_pdf

For more clearness some functions were added in 'IPreader'.

To get a possebilitiy for checking the PSL conversion a function 'plotscalefactors' was added.